requirements.txt
aiohttp==3.6.2
appdirs==1.4.3
APScheduler==3.6.3
async-timeout==3.0.1
attrs==20.2.0
CacheControl==0.12.6
certifi==2019.11.28
chardet==3.0.4
colorama==0.4.3
contextlib2==0.6.0
discord.py==1.5.0
distlib==0.3.0
distro==1.4.0
html5lib==1.0.1
idna==2.8
ipaddr==2.2.0
lockfile==0.12.2
msgpack==0.6.2
multidict==4.7.6
packaging==20.3
pep517==0.8.2
progress==1.5
pyparsing==2.4.6
pytoml==0.1.21
pytz==2020.1
requests==2.22.0
retrying==1.3.3
six==1.14.0
tzlocal==2.1
urllib3==1.25.8
webencodings==0.5.1
yarl==1.6.0
